sbicall

+++ kernel/lib/stdio.h
@@ -1,6 +1,7 @@
 #ifndef _STDIO_H
 #define _STDIO_H
+#include "../type/types.h"
 
-void putchar(int ch);
+void putchar(uint64 ch);
 
 #endif
